S2 migration done for now

+++ web-src/src/pages/Registrations/RegistrationsTab.tsx
@@ -3,14 +3,9 @@
 */
 
 import React, { useState, useCallback, useMemo } from 'react'
-import {
-  AlertDialog,
-  DialogTrigger as V3DialogTrigger,
-  ActionButton,
-} from '@adobe/react-spectrum'
-import { DialogTrigger, SearchField } from '@react-spectrum/s2'
+import { DialogTrigger, SearchField, ActionButton, AlertDialog } from '@react-spectrum/s2'
 import { style } from '@react-spectrum/s2/style' with { type: 'macro' }
-import Download from '@spectrum-icons/workflow/Download'
+import Download from '@react-spectrum/s2/icons/Download'
 import type { Attendee, AttendeeFilters, AttendeeColumnConfig } from '../../types/attendee'
 import type { Campaign } from '../../types/campaign'
 import { useHasPermission } from '../../hooks/useHasPermission'
@@ -195,30 +190,22 @@ export const RegistrationsTab: React.FC<RegistrationsTabProps> = ({
       </DialogTrigger>
 
       {/* Delete Confirmation Dialog */}
-      <V3DialogTrigger
-        isOpen={!!itemToDelete}
-        onOpenChange={(isOpen) => !isOpen && setItemToDelete(null)}
-      >
+      <DialogTrigger isOpen={!!itemToDelete} onOpenChange={(isOpen) => !isOpen && setItemToDelete(null)}>
         <div style={{ display: 'none' }} />
-        {(close) => (
-          <AlertDialog
-            title="Remove Attendee"
-            variant="destructive"
-            primaryActionLabel="Remove"
-            secondaryActionLabel="Cancel"
-            onPrimaryAction={() => {
-              if (itemToDelete) {
-                handleDeleteAttendee(itemToDelete)
-              }
-              close()
-            }}
-            onSecondaryAction={close}
-          >
-            Are you sure you want to remove {itemToDelete ? getAttendeeName(itemToDelete) : 'this attendee'}?
-            This action cannot be undone.
-          </AlertDialog>
-        )}
-      </V3DialogTrigger>
+        <AlertDialog
+          title="Remove Attendee"
+          variant="destructive"
+          primaryActionLabel="Remove"
+          cancelLabel="Cancel"
+          onPrimaryAction={() => {
+            if (itemToDelete) { handleDeleteAttendee(itemToDelete) }
+          }}
+          onCancel={() => setItemToDelete(null)}
+        >
+          Are you sure you want to remove {itemToDelete ? getAttendeeName(itemToDelete) : 'this attendee'}?
+          This action cannot be undone.
+        </AlertDialog>
+      </DialogTrigger>
     </>
   )
 }
